setting up delete functionality

+++ routes/ideas.js
@@ -46,16 +46,22 @@ router.get('/:id', async (req,res) => {
 router.put('/:id', async(req,res) => {
     try {
         const id = req.params.id
-        const updatedIdea = await Idea.findByIdAndUpdate(id, {
-            $set: {
-                text : req.body.text,
-                tag : req.body.tag,
-            }},
-            {
-                new: true
-            }
-        )
-        res.json({success: true, data: updatedIdea})
+        const idea = await Idea.findById(id)
+
+        if(idea.username !== req.body.username) {
+            res.status(403).json({success: true, error: 'User not authorized'})
+        } else {
+            const updatedIdea = await Idea.findByIdAndUpdate(id, {
+                $set: {
+                    text : req.body.text,
+                    tag : req.body.tag,
+                }},
+                {
+                    new: true
+                }
+            )
+            res.json({success: true, data: updatedIdea})
+        }
     } catch (error) {
         console.log(error)
         res.status(500).json({success: true, error: 'Something went wrong'})
@@ -67,8 +73,14 @@ router.put('/:id', async(req,res) => {
 router.delete('/:id', async(req,res) => {
     try {
         const id = req.params.id
-        await Idea.findByIdAndDelete(id)
-        res.json({success: true, data: {}})
+        const idea = await Idea.findById(id)
+
+        if(idea.username !== req.body.username) {
+            res.status(403).json({success: true, error: 'User not authorized'})
+        } else {
+            await Idea.findByIdAndDelete(id)
+            res.json({success: true, data: {}})
+        }
     } catch (error) {
         console.log(error)
         res.status(500).json({success: true, error: 'Something went wrong'})
